+++ contrib/parseq-benchmark/src/main/java/com/linkedin/parseq/HistogramSerializer.java
@@ -1,0 +1,11 @@
+package com.linkedin.parseq;
+
+import org.HdrHistogram.Histogram;
+
+public interface HistogramSerializer {
+
+  String serialize(Histogram histogram);
+
+  Histogram deserialize(String serialized);
+
+}