+++ test/com/linkedin/parseq/internal/TestSerialExecutor.java
@@ -14,10 +14,10 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import com.linkedin.parseq.internal.SerialExecutor.DeactivationListener;
-
 
 public class TestSerialExecutor {
   private ExecutorService _executorService;
@@ -25,14 +25,19 @@ public class TestSerialExecutor {
   private SerialExecutor _serialExecutor;
   private CapturingActivityListener _capturingDeactivationListener;
 
+  @DataProvider(name = "draining")
+  public Object[][] testDataProvider() {
+    return new Object[][] { { true }, { false } };
+  }
+
   @BeforeMethod
-  public void setUp() {
+  public void setUp(Object[] testArgs) {
     _executorService = new ThreadPoolExecutor(1, 1, 0, TimeUnit.SECONDS, new ArrayBlockingQueue<>(1),
         new ThreadPoolExecutor.AbortPolicy());
     _rejectionHandler = new CapturingExceptionHandler();
     _capturingDeactivationListener = new CapturingActivityListener();
     _serialExecutor = new SerialExecutor(_executorService, _rejectionHandler, _capturingDeactivationListener,
-        new FIFOPriorityQueue<>());
+        new FIFOPriorityQueue<>(), (boolean)testArgs[0]);
   }
 
   @AfterMethod
@@ -44,8 +49,8 @@ public class TestSerialExecutor {
     _executorService = null;
   }
 
-  @Test
-  public void testExecuteOneStepPlan() throws InterruptedException {
+  @Test(dataProvider = "draining")
+  public void testExecuteOneStepPlan(boolean draining) throws InterruptedException {
     final LatchedRunnable runnable = new LatchedRunnable();
     _serialExecutor.execute(runnable);
     assertTrue(runnable.await(5, TimeUnit.SECONDS));
@@ -54,8 +59,8 @@ public class TestSerialExecutor {
     assertEquals(_capturingDeactivationListener.getDeactivatedCount(), 1);
   }
 
-  @Test
-  public void testExecuteTwoStepPlan() throws InterruptedException {
+  @Test(dataProvider = "draining")
+  public void testExecuteTwoStepPlan(boolean draining) throws InterruptedException {
     final LatchedRunnable inner = new LatchedRunnable();
 
     final Runnable outer = new Runnable() {
@@ -72,8 +77,8 @@ public class TestSerialExecutor {
     assertEquals(_capturingDeactivationListener.getDeactivatedCount(), 1);
   }
 
-  @Test
-  public void testRejectOnFirstExecute() throws InterruptedException {
+  @Test(dataProvider = "draining")
+  public void testRejectOnFirstExecute(boolean draining) throws InterruptedException {
     // First fill up the underlying executor service so that a subsequent
     // submission of an execution loop by the serial executor will fail.
     _executorService.execute(new NeverEndingRunnable());
@@ -91,16 +96,18 @@ public class TestSerialExecutor {
         _rejectionHandler.getLastError() instanceof RejectedExecutionException);
   }
 
-  @Test
-  public void testRejectOnLoop() throws InterruptedException {
+  @Test(dataProvider = "draining")
+  public void testRejectOnLoop(boolean draining) throws InterruptedException {
     final CountDownLatch innerLatch = new CountDownLatch(1);
-    final LatchedRunnable innerRunnable = new LatchedRunnable();
+    final CountDownLatch outerLatch = new CountDownLatch(1);
+    final LatchedRunnable inner = new LatchedRunnable();
     final LatchedRunnable outer = new LatchedRunnable() {
       @Override
       public void run() {
         try {
+          outerLatch.countDown();
           innerLatch.await();
-          _serialExecutor.execute(innerRunnable);
+          _serialExecutor.execute(inner);
         } catch (InterruptedException e) {
           // Shouldn't happen
         }
@@ -111,26 +118,31 @@ public class TestSerialExecutor {
     // First we submit the outer task. This task will wait until for a count
     // down on the inner latch.
     _serialExecutor.execute(outer);
+    outerLatch.await();
 
     // Now we submit another runnable to the underlying executor. This should
     // get queued up.
     _executorService.execute(new NeverEndingRunnable());
 
     // Now the inner task will be submitted. This will eventually result in
-    // the re-execution of the serial executor loop, but that resubmission will
+    // the re-execution of the serial executor loop, but that re-submission will
     // fail because the thread is tied up by the current loop and the queue is
     // saturated with the previously submitted runnable.
     innerLatch.countDown();
 
-    assertTrue(_rejectionHandler.await(5, TimeUnit.SECONDS));
-    assertTrue(
-        "Expected " + _rejectionHandler.getLastError() + " to be instance of "
-            + RejectedExecutionException.class.getName(),
-        _rejectionHandler.getLastError() instanceof RejectedExecutionException);
+    if (!draining) {
+      assertTrue(_rejectionHandler.await(5, TimeUnit.SECONDS));
+      assertTrue(
+          "Expected " + _rejectionHandler.getLastError() + " to be instance of "
+              + RejectedExecutionException.class.getName(),
+          _rejectionHandler.getLastError() instanceof RejectedExecutionException);
+    } else {
+      assertFalse(_rejectionHandler.wasExecuted());
+    }
   }
 
-  @Test
-  public void testThrowingRunnable() throws InterruptedException {
+  @Test(dataProvider = "draining")
+  public void testThrowingRunnable(boolean draining) throws InterruptedException {
     _serialExecutor.execute(new ThrowingRunnable());
     assertTrue(_rejectionHandler.await(5, TimeUnit.SECONDS));
     assertTrue(