+++ test/com/linkedin/parseq/trace/TestTraceBuilder.java
@@ -45,7 +45,7 @@ public class TestTraceBuilder {
     assertEquals(trace2.build(), trace.getTraceMap().get(trace2.getId()));
     assertEquals(1, trace.getRelationships().size());
     assertTrue(trace.getRelationships()
-        .contains(new TraceRelationship(trace1.getId(), trace2.getId(), Relationship.SUCCESSOR_OF)));
+        .contains(new TraceRelationship(trace1, trace2, Relationship.SUCCESSOR_OF)));
   }
 
   @Test
@@ -62,19 +62,12 @@ public class TestTraceBuilder {
     assertEquals(trace2.build(), trace.getTraceMap().get(trace2.getId()));
     assertEquals(1, trace.getRelationships().size());
     assertTrue(trace.getRelationships()
-        .contains(new TraceRelationship(trace1.getId(), trace2.getId(), Relationship.SUCCESSOR_OF)));
+        .contains(new TraceRelationship(trace1, trace2, Relationship.SUCCESSOR_OF)));
   }
 
   @Test
   public void testRelationshipRetention() {
     final TraceBuilder builder = new TraceBuilder(4096, "test", 0L);
-    for (int i = 0; i < 4096 * 10; i++) {
-      final ShallowTraceBuilder trace1 =
-          new ShallowTraceBuilder(IdGenerator.getNextId()).setName("task1").setResultType(ResultType.UNFINISHED);
-      final ShallowTraceBuilder trace2 =
-          new ShallowTraceBuilder(IdGenerator.getNextId()).setName("task2").setResultType(ResultType.UNFINISHED);
-      builder.addRelationship(Relationship.SUCCESSOR_OF, trace1, trace2);
-    }
 
     List<TraceRelationship> rels = new ArrayList<TraceRelationship>();
     for (int i = 0; i < 4096; i++) {
@@ -83,7 +76,15 @@ public class TestTraceBuilder {
       final ShallowTraceBuilder trace2 =
           new ShallowTraceBuilder(IdGenerator.getNextId()).setName("task2").setResultType(ResultType.UNFINISHED);
       builder.addRelationship(Relationship.SUCCESSOR_OF, trace1, trace2);
-      rels.add(new TraceRelationship(trace1.getId(), trace2.getId(), Relationship.SUCCESSOR_OF));
+      rels.add(new TraceRelationship(trace1, trace2, Relationship.SUCCESSOR_OF));
+    }
+
+    for (int i = 0; i < 4096 * 10; i++) {
+      final ShallowTraceBuilder trace1 =
+          new ShallowTraceBuilder(IdGenerator.getNextId()).setName("task1").setResultType(ResultType.UNFINISHED);
+      final ShallowTraceBuilder trace2 =
+          new ShallowTraceBuilder(IdGenerator.getNextId()).setName("task2").setResultType(ResultType.UNFINISHED);
+      builder.addRelationship(Relationship.SUCCESSOR_OF, trace1, trace2);
     }
 
     Trace trace = builder.build();